speech-to-text-to-speech/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-speech>=2.31.1",
    "google-cloud-texttospeech>=2.25.1",
    "google-genai>=1.7.0",
    "python-dotenv>=1.0.1",
    "ruff>=0.11.2",
    "websockets>=15.0.1",
]
